Testing: Update test for babel-preset-default
