+++ packages/babel-preset-default/test/fixtures/input.js
@@ -1,0 +1,15 @@
+describe( 'Babel preset default', () => {
+	async function* foo() {
+		await 1;
+		yield 2;
+	}
+
+	test( 'support for async generator functions', async () => {
+		const generator = foo();
+
+		expect( await generator.next() ).toEqual( {
+			done: false,
+			value: 2,
+		} );
+	} );
+} );
